improvements are added

+++ src/main/java/com/att/orderservice/dto/OrderDto.java
@@ -1,7 +1,6 @@
 package com.att.orderservice.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.util.Date;
 
 public class OrderDto {
@@ -9,10 +8,10 @@ public class OrderDto {
     private Long id;
 
     @JsonProperty("user_id")
-    private Integer userId;
+    private Long userId;
 
     @JsonProperty("product_id")
-    private Integer productId;
+    private Long productId;
 
     private Integer quantity;
 
@@ -32,19 +31,19 @@ public class OrderDto {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(Integer productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
